Isolated CRUD operations for fetching a new schedule and saving a schedule into utils files
